Stage 2 of the Project - final polish
I added a settings activity.
The query is narrowed down by the preferences selected by the user for Number of articles to display and for the Section (topic) to display. The section topic has 5 options to choose (film, books, music, stage and games).
I struggled to limit the Number of articles input by user to maximum 50 articles (the response from the Guardian API is limited to 200!). I found that it was not so simple to do it.
I did it in:
- onPreferenceChange method where I had check and limit the input to 50 and than to write it to SharedPreferences file on device and in
- onCreateLoader method where I had checked what is retrieved from SharedPreferences file and limit it to 50.
It took me some extra time to do this :)